FT: rc: mysql.c: handle password SQL write failures Check f_write_string() results when creating the temporary setpasswd.sql file used during MySQL root password initialization.
On write failure, log the error, stop the temporary mysqld instance and
remove the temporary pid/sql files instead of continuing with incomplete
setup data.

This prevents a failed or partial SQL file write from being treated as a
successful password initialization, especially on low-space or failing
storage.

Update from FreshTomato
https://github.com/FreshTomato-Project/freshtomato-arm/commit/38cdd1a12ca3b929a595e34246438e3834125e18

+++ tomato64/package/rc/rc/mysql.c
@@ -501,14 +501,13 @@ void start_mysql(int force)
 		rc = mysql_to_hex(pass_hex, sizeof(pass_hex), nvram_safe_get("mysql_passwd"));
 		if (rc != 0) {
 			logmsg(LOG_ERR, "%s: mysql password is too long", __FUNCTION__);
+			killall_tk_period_wait("mysqld", 50);
+			sleep(1);
+			eval("rm", "-f", mysql_pid, mysql_passwd);
 			goto END;
 		}
 
-		if (f_exists(mysql_passwd))
-			unlink(mysql_passwd);
-
 #ifndef TOMATO64
-		f_write_string(mysql_passwd, "use mysql;", FW_CREATE | FW_NEWLINE, 0600);
 		if (pass_hex[0])
 			rc = snprintf(sql, sizeof(sql), "update user set password=password(0x%s) where user='root';", pass_hex);
 		else
@@ -517,7 +516,6 @@ void start_mysql(int force)
 		/* MariaDB 10.4+ dropped the plain `password` column on mysql.user;
 		 * use the plugin-aware SET PASSWORD statement instead.
 		 */
-		f_write_string(mysql_passwd, "flush privileges;", FW_CREATE | FW_NEWLINE, 0600);
 		if (pass_hex[0])
 			rc = snprintf(sql, sizeof(sql), "SET PASSWORD FOR 'root'@'localhost' = PASSWORD(0x%s);", pass_hex);
 		else
@@ -525,16 +523,38 @@ void start_mysql(int force)
 #endif /* TOMATO64 */
 		if ((rc < 0) || (rc >= (int)sizeof(sql))) {
 			logmsg(LOG_ERR, "%s: password SQL is too long", __FUNCTION__);
-			unlink(mysql_passwd);
+			killall_tk_period_wait("mysqld", 50);
+			sleep(1);
+			eval("rm", "-f", mysql_pid, mysql_passwd);
 			goto END;
 		}
-		f_write_string(mysql_passwd, sql, FW_APPEND | FW_NEWLINE, 0);
-		f_write_string(mysql_passwd, "flush privileges;", FW_APPEND | FW_NEWLINE, 0);
+
+		if (f_exists(mysql_passwd))
+			unlink(mysql_passwd);
+
+#ifndef TOMATO64
+		if ((f_write_string(mysql_passwd, "use mysql;", FW_CREATE | FW_NEWLINE, 0600) < 0) ||
+		    (f_write_string(mysql_passwd, sql, FW_APPEND | FW_NEWLINE, 0) < 0) ||
+		    (f_write_string(mysql_passwd, "flush privileges;", FW_APPEND | FW_NEWLINE, 0) < 0)) {
+#else
+		/* SET PASSWORD operates at server scope, so start with a flush instead of `use mysql;`. */
+		if ((f_write_string(mysql_passwd, "flush privileges;", FW_CREATE | FW_NEWLINE, 0600) < 0) ||
+		    (f_write_string(mysql_passwd, sql, FW_APPEND | FW_NEWLINE, 0) < 0) ||
+		    (f_write_string(mysql_passwd, "flush privileges;", FW_APPEND | FW_NEWLINE, 0) < 0)) {
+#endif /* TOMATO64 */
+			logerr(__FUNCTION__, __LINE__, mysql_passwd);
+			killall_tk_period_wait("mysqld", 50);
+			sleep(1);
+			eval("rm", "-f", mysql_pid, mysql_passwd);
+			goto END;
+		}
 
 		rc = snprintf(sql, sizeof(sql), "source %s", mysql_passwd);
 		if ((rc < 0) || (rc >= (int)sizeof(sql))) {
 			logmsg(LOG_ERR, "%s: mysql source command is too long", __FUNCTION__);
-			unlink(mysql_passwd);
+			killall_tk_period_wait("mysqld", 50);
+			sleep(1);
+			eval("rm", "-f", mysql_pid, mysql_passwd);
 			goto END;
 		}
